remove otel test duplication
