+++ packages/ai/test/vercel/dual-version.test.ts
@@ -1,41 +1,24 @@
 import { describe, it, expect, beforeEach, beforeAll, afterAll, vi } from 'vitest';
-import { InMemorySpanExporter, SimpleSpanProcessor } from '@opentelemetry/sdk-trace-base';
-import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
-import { trace } from '@opentelemetry/api';
 import { wrapAISDKModel } from '../../src/otel/vercel';
 import { createMockProvider } from './mock-provider-v1/mock-provider';
 import { createMockProvider as createMockProviderV2 } from './mock-provider-v2/mock-provider-v2';
 import { generateText } from 'aiv4';
 import { generateText as generateTextV5 } from 'aiv5';
 import { withSpan } from '../../src/otel/withSpan';
-import { initAxiomAI, resetAxiomAI } from '../../src/otel/initAxiomAI';
+import { createOtelTestSetup } from '../helpers/otel-test-setup';
 
-let memoryExporter: InMemorySpanExporter;
-let tracerProvider: NodeTracerProvider;
+const otelTestSetup = createOtelTestSetup();
 
 beforeAll(() => {
-  memoryExporter = new InMemorySpanExporter();
-  const spanProcessor = new SimpleSpanProcessor(memoryExporter);
-  tracerProvider = new NodeTracerProvider({
-    spanProcessors: [spanProcessor],
-  });
-  tracerProvider.register();
-
-  // Initialize AxiomAI with the tracer to prevent "No tracer found" warnings
-  const tracer = trace.getTracer('axiom-ai-test');
-  initAxiomAI({ tracer });
+  otelTestSetup.setup();
 });
 
 beforeEach(() => {
-  memoryExporter.reset();
+  otelTestSetup.reset();
 });
 
 afterAll(async () => {
-  // Reset AxiomAI configuration before shutting down
-  resetAxiomAI();
-
-  await tracerProvider.shutdown();
-  await memoryExporter.shutdown();
+  await otelTestSetup.cleanup();
 });
 
 describe('Dual Version Support', () => {
@@ -60,7 +43,7 @@ describe('Dual Version Support', () => {
       expect(result.text).toBe('Hello from v1!');
       expect(model.specificationVersion).toBe('v1');
 
-      const spans = memoryExporter.getFinishedSpans();
+      const spans = otelTestSetup.getSpans();
       expect(spans).toHaveLength(1);
       expect(spans[0].name).toBe('chat test-model');
       expect(spans[0].attributes['gen_ai.request.model']).toBe('test-model');
@@ -86,7 +69,7 @@ describe('Dual Version Support', () => {
       expect(result.text).toBe('Hello from v2!');
       expect(model.specificationVersion).toBe('v2');
 
-      const spans = memoryExporter.getFinishedSpans();
+      const spans = otelTestSetup.getSpans();
       expect(spans).toHaveLength(1);
       expect(spans[0].name).toBe('chat test-model-v2');
       expect(spans[0].attributes['gen_ai.request.model']).toBe('test-model-v2');
@@ -128,7 +111,7 @@ describe('Dual Version Support', () => {
         });
       });
 
-      const spans = memoryExporter.getFinishedSpans();
+      const spans = otelTestSetup.getSpans();
       const span = spans[0];
 
       expect(span.attributes['gen_ai.usage.input_tokens']).toBe(20);
@@ -152,7 +135,7 @@ describe('Dual Version Support', () => {
         });
       });
 
-      const spans = memoryExporter.getFinishedSpans();
+      const spans = otelTestSetup.getSpans();
       const span = spans[0];
 
       expect(span.attributes['gen_ai.usage.input_tokens']).toBe(25);
@@ -178,7 +161,7 @@ describe('Dual Version Support', () => {
         });
       });
 
-      const spans = memoryExporter.getFinishedSpans();
+      const spans = otelTestSetup.getSpans();
       expect(spans).toHaveLength(1);
       expect(spans[0].attributes['gen_ai.response.finish_reasons']).toBe('["length"]');
     });
@@ -200,7 +183,7 @@ describe('Dual Version Support', () => {
         });
       });
 
-      const spans = memoryExporter.getFinishedSpans();
+      const spans = otelTestSetup.getSpans();
       expect(spans).toHaveLength(1);
       expect(spans[0].attributes['gen_ai.response.finish_reasons']).toBe('["length"]');
     });
