+++ packages/ai/test/vercel/features.test.ts
@@ -1,39 +1,23 @@
 import { describe, expect, it, beforeAll, beforeEach, afterAll } from 'vitest';
-import { InMemorySpanExporter, SimpleSpanProcessor } from '@opentelemetry/sdk-trace-base';
-import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import { wrapAISDKModel } from '../../src/otel/vercel';
 import { withSpan } from '../../src/otel/withSpan';
 import { generateText } from 'aiv4';
 import { createMockProvider, mockResponses } from './mock-provider-v1/mock-provider';
-import { SpanKind, trace } from '@opentelemetry/api';
-import { initAxiomAI, resetAxiomAI } from '../../src/otel/initAxiomAI';
+import { SpanKind } from '@opentelemetry/api';
+import { createOtelTestSetup } from '../helpers/otel-test-setup';
 
-let memoryExporter: InMemorySpanExporter;
-let tracerProvider: NodeTracerProvider;
+const otelTestSetup = createOtelTestSetup();
 
 beforeAll(() => {
-  memoryExporter = new InMemorySpanExporter();
-  const spanProcessor = new SimpleSpanProcessor(memoryExporter);
-  tracerProvider = new NodeTracerProvider({
-    spanProcessors: [spanProcessor],
-  });
-  tracerProvider.register();
-
-  // Initialize AxiomAI with the tracer to prevent "No tracer found" warnings
-  const tracer = trace.getTracer('axiom-ai-test');
-  initAxiomAI({ tracer });
+  otelTestSetup.setup();
 });
 
 beforeEach(() => {
-  memoryExporter.reset();
+  otelTestSetup.reset();
 });
 
 afterAll(async () => {
-  // Reset AxiomAI configuration before shutting down
-  resetAxiomAI();
-
-  await tracerProvider.shutdown();
-  await memoryExporter.shutdown();
+  await otelTestSetup.cleanup();
 });
 
 describe('span names', () => {
@@ -49,7 +33,7 @@ describe('span names', () => {
       });
     });
 
-    const spans = memoryExporter.getFinishedSpans();
+    const spans = otelTestSetup.getSpans();
     expect(spans.length).toBe(1);
     expect(spans[0].name).toBe('chat model-name');
   });
@@ -67,7 +51,7 @@ describe('span names', () => {
       });
     });
 
-    const spans = memoryExporter.getFinishedSpans();
+    const spans = otelTestSetup.getSpans();
     expect(spans[0].kind).toBe(SpanKind.INTERNAL);
   });
 
@@ -85,7 +69,7 @@ describe('span names', () => {
       });
     });
 
-    const spans = memoryExporter.getFinishedSpans();
+    const spans = otelTestSetup.getSpans();
     expect(spans[0].attributes.foo).toBe('bar');
   });
 });
